Updating constants to be in a difference file and commenting out test scraper call
