+++ scrapers/src/catalog-scraper/scraper_constants.ts
@@ -1,0 +1,407 @@
+// sections that need to be overrided as AND sections
+export const ANDSectionHeader: string[] = [
+  "Intermediate and Advanced Biology Electives",
+  "History Electives",
+  // "Sociology Electives",
+];
+
+// Headers of sections that should be parsed as Subheader sections
+export const SubheaderTagSet: string[] = [
+  "Choose one of the two options.",
+  "Complete two courses for one of the following science categories:",
+  "Students are required to complete one of the following foci (two courses total):",
+];
+
+// Dictionary for ORSection keywords, maps from keyword/phrase -> number of credits
+export const ORTagMap: { [key: string]: number } = {
+  "Then complete one of the following:": 4,
+  "Choose one of the following:": 4,
+  "Complete one of the following:": 4,
+  "Complete one of the following: ": 4,
+  "Complete one course from the following:": 4,
+  "Complete two of the following:": 8,
+  "Complete three of the following:": 12,
+  "Complete four of the following:": 16,
+  "Complete six of the following:": 24,
+  "Complete one of the following sequences:": 10,
+  "Complete two courses for one of the following science categories:": 10,
+  "Complete five courses from the following:": 20,
+  "Complete two courses from the following lists:": 8,
+  "Students are required to complete one of the following foci (two courses total):": 8,
+  "Complete one of the following courses. This course may also be used to fulfill an additional English requirement below:": 4,
+  "Complete at least two of the following:": 8,
+  "Complete four economics electives with no more than two below 3000:": 16,
+  "Complete one course from one of the following groups:": 4,
+  "Take two courses, at least one of which is at the 4000 or 5000 level, from the following:": 8,
+  "Complete one of the following courses not already taken:": 4,
+  "Complete two of the following courses not already taken:": 8,
+  "Complete two courses from the following:": 8,
+  "Complete two biology courses (with corequisite labs if offered). Choose one of these two courses from the following list:": 8,
+  "Complete one introductory course from the following:": 4,
+  "Complete one capstone experience from the following:": 4,
+  "Complete three courses from the following:": 12,
+  "Complete either one computer science capstone or the physics capstone:": 4,
+  "Complete one of the following lecture/lab pairs. PHYS 1145/PHYS 1146 is recommended:": 4,
+  "Choose one:": 4,
+  "Choose one": 4,
+  "Complete four courses from the following:": 16,
+  "Complete any six courses as long as prerequisites have been met. At least two must be a 3000-level course. ": 24,
+  "Choose one of the two options.": 12,
+  // TODO: Data-Science-Related Electives: "Complete six courses from categories A and B, at least three of which must be from B"
+};
+
+// Tags for sections that should be ignored entirely
+export const IgnoreTags: string[] = [
+  "The following courses are used in other areas of the major:",
+  "The following courses are fulfilled through the computer science requirement:",
+  "Note: The following courses do not count toward this concentration:",
+];
+
+// Dictionary for RANGESection keywords, maps from keyword/phrase -> number of credits
+export const RANGETagMap: { [key: string]: number } = {
+  "Complete 4 credits of CS, IS, or DS classes that are not already required. Choose courses within the following ranges:": 4,
+  "Complete 4 credits of CS, IS, or DS courses that are not already required. Choose courses within the following range:": 4,
+  "Complete four credits of CS, IS, or DS classes that are not already required. Choose courses within the following ranges:": 4,
+  "Complete 8 credits of CS, IS or DS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete 8 credits of CS, IS or, DS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete 8 credits of CS, IS, or DS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete 8 credits of upper-division CS, IS, or DS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete eight credits of CS, IS or, DS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete eight credits of CS, IS or DS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete 12 credits of CS, IS or, DS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 12 credits of CS, IS, or DS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 12 credits of upper-division CS, IS, and DS courses that are not already required. Choose courses within the following ranges:": 12,
+  "Complete twelve credits of CS, IS or DS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 16 credits of CS, IS, or DS classes that are not already required. Choose courses within the following ranges:": 16,
+  "Complete 16 credits of upper-division CS, IS, or DS courses that are not already required. Choose courses within the following ranges:": 16,
+
+  "Complete one of the following, not taken to fulfill previous requirements:": 4,
+  "Complete one from the following:": 4,
+  "Complete four credits from the following:": 16,
+  "Complete four courses in the following range:": 16,
+  "Complete three courses in the following range:": 12,
+
+  "Complete four ECON electives with at least two numbered at ECON 3000 or above.": 16,
+
+  "Choose the second elective from the following list:": 4,
+  "Complete three intermediate/advanced-level courses:": 12,
+  "Intermediate/Advanced History Cluster ": 12,
+  "Complete one advanced-level course:": 4,
+  // "Complete one sociology elective in each of the following ranges:": 4,
+
+  "Complete four credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 4,
+  "Complete 4 credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 4,
+  "Complete 4 credits of CS, CY, DS, or IS courses that are not already required. Choose courses within the following range:": 4,
+  "Complete 8 credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete 8 credits of CS, CY, DS or IS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete 8 credits of upper-division CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete eight credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 8,
+  "Complete twelve credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 12 credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 12 credits of upper-division CS, CY, DS, or IS courses that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 12 credits of upper-division CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Comment: Complete 12 credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 12,
+  "Complete 16 credits of upper-division CS, CY, DS, or IS courses that are not already required. Choose courses within the following ranges:": 16,
+  "Complete one course from the following": 4,
+  "Complete four credits from the following list, not taken to fulfill previous requirements:": 4,
+  "Complete 4 Economics elective courses from the following ranges with no more than 2 at the ECON 1200 – ECON 1999 range:": 4,
+  "Complete 4 ECON electives that are found in the following ranges, with at least two numbered at ECON 3000 or above:": 16,
+  "Complete two courses in the following range:": 8,
+
+  "Complete four credits of CS,CY, DS, or IS classes that are not already required. Choose courses within the following ranges:": 16,
+  "Complete one course not already required in the following range:": 4,
+  "Complete 4 credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:	": 4,
+  "Complete four credits of CS, CY, DS, or IS classes that are not already required. Choose courses within the following ranges:	": 4,
+};
+
+// Set for RANGESections that only indicate the major in which they are allowed to take electives
+export const RANGECourseSet: Array<string> = [
+  "ARTD",
+  "ARTE",
+  "ARTF",
+  "ARTG",
+  "ARTH",
+  "ARTS",
+  "CRIM",
+  "ENGL",
+  "GAME",
+  "HIST",
+  "JRNL",
+  "PHIL",
+];
+
+// All valid subject tags for parsing courses
+export const ValidSubjects: string[] = [
+  "ACCT",
+  "ACC",
+  "AFAM",
+  "AFRS",
+  "AMSL",
+  "ASL",
+  "ALY",
+  "ANTH",
+  "ANT",
+  "ARAB",
+  "ARCH",
+  "ARMY",
+  "ART",
+  "ARTG",
+  "ARTF",
+  "ARTE",
+  "ARTH",
+  "ARTD",
+  "AACE",
+  "ARTS",
+  "ASNS",
+  "BNSC",
+  "BIOC",
+  "BIOE",
+  "BINF",
+  "BIOL",
+  "BIO",
+  "BIOT",
+  "BTC",
+  "BUSN",
+  "EXSC",
+  "CHME",
+  "CHEM",
+  "CHM",
+  "CHMY",
+  "CHNS",
+  "CIVE",
+  "CED",
+  "COMM",
+  "CMN",
+  "CMMN",
+  "CET",
+  "ETC",
+  "CS",
+  "CSYE",
+  "CMG",
+  "COOP",
+  "COP",
+  "EEAM",
+  "EESC",
+  "EESH",
+  "CAEP",
+  "CRS",
+  "CJS",
+  "CRIM",
+  "CLTR",
+  "LITR",
+  "CY",
+
+  "DA",
+  "DS",
+  "DEAF",
+  "DGM",
+
+  "GEO",
+  "EVRN",
+  "ENVR",
+  "ESC",
+  "EEMB",
+  "ECN",
+  "ECNM",
+  "ECON",
+  "EDU",
+  "EDUC",
+  "EECE",
+  "EET",
+  "ENSY",
+  "ENCP",
+  "ENGR",
+  "ENLR",
+  "EMGT",
+  "ESLG",
+  "ESL",
+  "ENG",
+  "ENGH",
+  "ENGL",
+  "ENGW",
+  "EAI",
+  "ENTR",
+  "TECE",
+  "ENVS",
+  "FINA",
+  "FIN",
+  "FSEM",
+  "FLNG",
+  "FRN",
+  "FRNH",
+  "GAME",
+  "GSND",
+  "GEN",
+  "GENR",
+  "GE",
+  "GET",
+  "GENS",
+  "GIS",
+  "GRMN",
+  "GST",
+  "GBST",
+  "GSE",
+  "GREK",
+  "HINF",
+  "HIA",
+  "HMG",
+  "HSC",
+  "HSCI",
+  "HLTH",
+  "HBRW",
+  "HST",
+  "HSTY",
+  "HIST",
+  "HLS",
+  "HONR",
+  "HPA",
+  "HUM",
+  "HRM",
+  "HRMG",
+  "HSV",
+  "HUSV",
+
+  "IE",
+  "IA",
+  "IRM",
+  "IS",
+  "INFO",
+  "ITC",
+  "ITP",
+  "INT",
+  "INPS",
+  "INAM",
+  "INSC",
+  "INSH",
+  "INPR",
+  "INTL",
+  "INTB",
+  "INTP",
+  "ITLN",
+
+  "JPNS",
+  "JLS",
+  "JWSS",
+  "JRNL",
+
+  "KMG",
+
+  "LARC",
+  "LANG",
+  "LACS",
+  "LWP",
+  "LPSC",
+  "LW",
+  "LAW",
+  "LDR",
+  "LS",
+  "LST",
+  "LING",
+
+  "MGT",
+  "MIS",
+  "MISM",
+  "MGMT",
+  "MNSC",
+  "MGSC",
+  "MECN",
+  "MARS",
+  "MKT",
+  "MKTG",
+  "MATL",
+  "MTH",
+  "MATM",
+  "MATH",
+  "MEIE",
+  "ME",
+  "MET",
+  "MSCR",
+  "CINE",
+  "LNG",
+  "MUS",
+  "MSIC",
+  "MUSI",
+  "MUSC",
+  "MPNC",
+  "MUST",
+
+  "NNMD",
+  "NAVY",
+  "NETS",
+  "NPM",
+  "NRSG",
+  "NTR",
+
+  "OR",
+  "ORGB",
+  "OCM",
+
+  "PHSC",
+  "PMST",
+  "PMCL",
+  "PMC",
+  "PHMD",
+  "PHDL",
+  "PHL",
+  "PHLS",
+  "PHIL",
+  "RELS",
+  "PHE",
+  "PTH",
+  "PT",
+  "PA",
+  "PHY",
+  "PHYC",
+  "PHYS",
+  "POL",
+  "PLSC",
+  "POLS",
+  "PORT",
+  "PJM",
+  "PSY",
+  "PSYC",
+  "PHTH",
+  "PPUA",
+  "PBR",
+
+  "RGA",
+  "RFA",
+  "RMS",
+  "RPT",
+  "RSSN",
+
+  "SMFA",
+  "SCI",
+  "SOC",
+  "SOCL",
+  "SPN",
+  "SPNS",
+  "SSAM",
+  "SLPA",
+  "SIA",
+  "STRT",
+  "ABRD",
+  "ABRB",
+  "ABRC",
+  "ABRL",
+  "ABRS",
+  "ABRH",
+  "ABRU",
+  "SCHM",
+  "SBSY",
+  "SUEN",
+  "SWHL",
+  "TCC",
+  "TCM",
+  "TELR",
+  "TELE",
+  "THTR",
+  "TOXC",
+  "URBS",
+  "WMNS",
+];
+
+// Majors that have optional concentrations and therefore should have a min concentration value of 0
+export const OPTIONAL_CONCENTRATION = [
+  "Computer Science and Political Science, BS",
+];
